change data type

+++ database/migrations/2016_04_29_004613_create_order_details_table.php
@@ -17,7 +17,7 @@ class CreateOrderDetailsTable extends Migration
             $table->string('product_link');
             $table->unsignedInteger('quantity');
             $table->unsignedInteger('unit_price');
-            $table->unsignedInteger('weight');
+            $table->decimal('weight');
             $table->unsignedInteger('order_id');
             $table->timestamps();
             $table->foreign('order_id')->references('id')->on('orders')->onDelete('cascade');
